+++ src/calendarjs.js
@@ -324,6 +324,7 @@ function calendarJs( id, options, startDateTime ) {
         _element_SelectExportTypeDialog_Option_JSON = null,
         _element_SelectExportTypeDialog_Option_TEXT = null,
         _element_SelectExportTypeDialog_Option_iCAL = null,
+        _element_SelectExportTypeDialog_Option_MD = null,
         _element_SelectExportTypeDialog_ExportEvents = null,
         _element_Tooltip = null,
         _element_Tooltip_Title = null,
@@ -3310,6 +3311,7 @@ function calendarJs( id, options, startDateTime ) {
         _element_SelectExportTypeDialog_Option_JSON = buildRadioButton( radioButtonsContainer, "JSON", "ExportType" );
         _element_SelectExportTypeDialog_Option_TEXT = buildRadioButton( radioButtonsContainer, "TEXT", "ExportType" );
         _element_SelectExportTypeDialog_Option_iCAL = buildRadioButton( radioButtonsContainer, "iCAL", "ExportType" );
+        _element_SelectExportTypeDialog_Option_MD = buildRadioButton( radioButtonsContainer, "MD", "ExportType" );
 
         var buttonsSplitContainer = createElement( "div", "split" );
         contents.appendChild( buttonsSplitContainer );
@@ -3350,6 +3352,8 @@ function calendarJs( id, options, startDateTime ) {
             exportEvents( _element_SelectExportTypeDialog_ExportEvents, "text" );
         } else if ( _element_SelectExportTypeDialog_Option_iCAL.checked ) {
             exportEvents( _element_SelectExportTypeDialog_ExportEvents, "ical" );
+        } else if ( _element_SelectExportTypeDialog_Option_MD.checked ) {
+            exportEvents( _element_SelectExportTypeDialog_ExportEvents, "md" );
         }
     }
 
@@ -4183,6 +4187,8 @@ function calendarJs( id, options, startDateTime ) {
             contents = getTextContents( contentsEvents );
         } else if ( type === "ical" ) {
             contents = getICalContents( contentsEvents );
+        } else if ( type === "md" ) {
+            contents = getMdContents( contentsEvents );
         }
 
         if ( contents !== "" ) {
@@ -4199,6 +4205,8 @@ function calendarJs( id, options, startDateTime ) {
                 extension = "ics";
             } else if ( type === "json" ) {
                 mimeTypeStart = "application";
+            } else if ( type === "md" ) {
+                mimeTypeEnd = "x-markdown";
             }
 
             tempLink.style.display = "none";
@@ -4358,6 +4366,55 @@ function calendarJs( id, options, startDateTime ) {
         return getArrayText( daysNames );
     }
 
+    function getExportHeaders() {
+        var headers = [ 
+                _options.idText,
+                _options.fromText,
+                _options.toText,
+                _options.isAllDayText,
+                _options.titleText,
+                _options.descriptionText,
+                _options.locationText,
+                _options.backgroundColorText,
+                _options.textColorText,
+                _options.borderColorText,
+                _options.repeatsText,
+                _options.repeatEndsText,
+                _options.repeatDaysToExcludeText,
+                _options.seriesIgnoreDatesText,
+                _options.createdText,
+                _options.organizerNameText,
+                _options.organizerEmailAddressText
+            ],
+            headersLength = headers.length;
+
+        return [ headers, headersLength ];
+    }
+
+    function getExportRow( eventDetails ) {
+        var eventContents = [];
+
+        eventContents.push( getString( eventDetails.id ) );
+        eventContents.push( getStringFromDateTime( eventDetails.from ) );
+        eventContents.push( getStringFromDateTime( eventDetails.to ) );
+        eventContents.push( getYesNoFromBoolean( eventDetails.isAllDay ) );
+        eventContents.push( getString( eventDetails.title ) );
+        eventContents.push( getString( eventDetails.description ) );
+        eventContents.push( getString( eventDetails.location ) );
+        eventContents.push( getString( eventDetails.color ) );
+        eventContents.push( getString( eventDetails.colorText ) );
+        eventContents.push( getString( eventDetails.colorBorder ) );
+        eventContents.push( getRepeatsText( eventDetails.repeatEvery ) );
+        eventContents.push( getStringFromDateTime( eventDetails.repeatEnds ) );
+        eventContents.push( getArrayDays( eventDetails.repeatEveryExcludeDays ) );
+        eventContents.push( getArrayText( eventDetails.seriesIgnoreDates ) );
+        eventContents.push( getStringFromDateTime( eventDetails.created ) );
+        eventContents.push( getString( eventDetails.organizerName ) );
+        eventContents.push( getString( eventDetails.organizerEmailAddress ) );
+
+        return eventContents;
+    }
+
 
     /*
      * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
@@ -4366,8 +4423,10 @@ function calendarJs( id, options, startDateTime ) {
      */
 
     function getCsvContents( orderedEvents ) {
-        var headers = [ _options.idText, _options.fromText, _options.toText, _options.isAllDayText, _options.titleText, _options.descriptionText, _options.locationText, _options.backgroundColorText, _options.textColorText, _options.borderColorText, _options.repeatsText, _options.repeatEndsText, _options.repeatDaysToExcludeText, _options.seriesIgnoreDatesText, _options.createdText, _options.organizerNameText, _options.organizerEmailAddressText ],
-            headersLength = headers.length,
+        var orderedEventLength = orderedEvents.length,
+            exportHeaders = getExportHeaders(),
+            headers = exportHeaders[ 0 ],
+            headersLength = exportHeaders[ 1 ],
             csvHeaders = [],
             csvContents = [];
 
@@ -4377,7 +4436,6 @@ function calendarJs( id, options, startDateTime ) {
         
         csvContents.push( getCsvValueLine( csvHeaders ) );
 
-        var orderedEventLength = orderedEvents.length;
         for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventLength; orderedEventIndex++ ) {
             storeCsvData( csvContents, orderedEvents[ orderedEventIndex ] );
         }
@@ -4386,25 +4444,12 @@ function calendarJs( id, options, startDateTime ) {
     }
 
     function storeCsvData( csvContents, eventDetails ) {
-        var eventContents = [];
+        var eventContents = getExportRow( eventDetails ),
+            eventContentsLength = eventContents.length;
 
-        eventContents.push( getCsvValue( getString( eventDetails.id ) ) );
-        eventContents.push( getCsvValue( getStringFromDateTime( eventDetails.from ) ) );
-        eventContents.push( getCsvValue( getStringFromDateTime( eventDetails.to ) ) );
-        eventContents.push( getCsvValue( getYesNoFromBoolean( eventDetails.isAllDay ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.title ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.description ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.location ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.color ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.colorText ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.colorBorder ) ) );
-        eventContents.push( getCsvValue( getRepeatsText( eventDetails.repeatEvery ) ) );
-        eventContents.push( getCsvValue( getStringFromDateTime( eventDetails.repeatEnds ) ) );
-        eventContents.push( getCsvValue( getArrayDays( eventDetails.repeatEveryExcludeDays ) ) );
-        eventContents.push( getCsvValue( getArrayText( eventDetails.seriesIgnoreDates ) ) );
-        eventContents.push( getCsvValue( getStringFromDateTime( eventDetails.created ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.organizerName ) ) );
-        eventContents.push( getCsvValue( getString( eventDetails.organizerEmailAddress ) ) );
+        for ( var eventContentsIndex = 0; eventContentsIndex < eventContentsLength; eventContentsIndex++ ) {
+            eventContents[ eventContentsIndex ] = getCsvValue( eventContents[ eventContentsIndex ] );
+        }
 
         csvContents.push( getCsvValueLine( eventContents ) );
     }
@@ -4429,11 +4474,12 @@ function calendarJs( id, options, startDateTime ) {
      */
 
     function getXmlContents( orderedEvents ) {
-        var contents = [];
+        var contents = [],
+            orderedEventLength = orderedEvents.length;
+
         contents.push( "<?xml version=\"1.0\" ?>" );
         contents.push( "<Events>" );
 
-        var orderedEventLength = orderedEvents.length;
         for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventLength; orderedEventIndex++ ) {
             var orderedEvent = orderedEvents[ orderedEventIndex ];
 
@@ -4463,11 +4509,12 @@ function calendarJs( id, options, startDateTime ) {
      */
 
     function getJsonContents( orderedEvents ) {
-        var contents = [];
+        var contents = [],
+            orderedEventLength = orderedEvents.length;
+
         contents.push( "{" );
         contents.push( "\"events:\": [" );
 
-        var orderedEventLength = orderedEvents.length;
         for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventLength; orderedEventIndex++ ) {
             var orderedEvent = orderedEvents[ orderedEventIndex ];
             
@@ -4499,9 +4546,9 @@ function calendarJs( id, options, startDateTime ) {
      */
 
     function getTextContents( orderedEvents ) {
-        var contents = [];
+        var contents = [],
+            orderedEventLength = orderedEvents.length;
 
-        var orderedEventLength = orderedEvents.length;
         for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventLength; orderedEventIndex++ ) {
             var orderedEvent = orderedEvents[ orderedEventIndex ];
 
@@ -4527,12 +4574,13 @@ function calendarJs( id, options, startDateTime ) {
      */
 
     function getICalContents( orderedEvents ) {
-        var contents = [];
+        var contents = [],
+            orderedEventLength = orderedEvents.length;
+
         contents.push( "BEGIN:VCALENDAR" );
         contents.push( "VERSION:2.0" );
         contents.push( "PRODID:-//hacksw/handcal//NONSGML v1.0//EN" );
 
-        var orderedEventLength = orderedEvents.length;
         for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventLength; orderedEventIndex++ ) {
             var orderedEvent = orderedEvents[ orderedEventIndex ];
 
@@ -4547,6 +4595,35 @@ function calendarJs( id, options, startDateTime ) {
         }
 
         contents.push( "END:VCALENDAR" );
+
+        return contents.join( "\n" );
+    }
+
+
+    /*
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     * Exporting To MD
+     * ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+     */
+
+    function getMdContents( orderedEvents ) {
+        var orderedEventLength = orderedEvents.length,
+            exportHeaders = getExportHeaders(),
+            headersLength = exportHeaders[ 1 ],
+            contents = [ "| " + exportHeaders[ 0 ].join( " | " ) + " |" ],
+            headerLines = [];
+
+        for ( var headerIndex = 0; headerIndex < headersLength; headerIndex++ ) {
+            headerLines.push( "---" );
+        }
+
+        contents.push( "| " + headerLines.join( " | " ) + " |" );
+
+        for ( var orderedEventIndex = 0; orderedEventIndex < orderedEventLength; orderedEventIndex++ ) {
+            var rowContents = getExportRow( orderedEvents[ orderedEventIndex ] );
+
+            contents.push( "| " + rowContents.join( " | " ) + " |" );
+        }
 
         return contents.join( "\n" );
     }
@@ -4698,7 +4775,7 @@ function calendarJs( id, options, startDateTime ) {
      * 
      * @fires onEventsExported
      * 
-     * @param       {string}    type                                        The data type to export to (defaults to "csv", accepts "csv", "xml", "json", "txt" and "ical").
+     * @param       {string}    type                                        The data type to export to (defaults to "csv", accepts "csv", "xml", "json", "txt", "ical", and "md").
      */
     this.exportAllEvents = function( type ) {
         if ( _options.exportEventsEnabled ) {